[#3084] Move MqttDeviceEndpoint class to upper level.
